Add backend cache and implement cbf on the backend

+++ backend/src/models/user.rs
@@ -1,5 +1,5 @@
 pub struct User {
-    pub user_id: i32,
+    pub user_id: i16,
     pub username: String,
     pub email: String,
     pub password: [u8; 48], // 16 bytes for the salt and 32 bytes for the hash
